+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v14.0.0
 
 - **_Breaking Change_**: remove `AdvancedEvent`, add `UniEvent`, the interface is the same and the performance is better.
+- **_Breaking Change_**: rename `IReadonlyIoC` to `IReadOnlyIoCC`.
 
 ## v13.1.0